Move Token into it's own file
